+++ src/lib/utils/timezones.ts
@@ -1,51 +1,158 @@
 import tzlookup from 'tz-lookup'
-import { DateTime } from 'luxon'
+import { IANAZone, DateTime } from 'luxon'
+import { logr } from './logr'
+import { CmfEvent } from '@/types/events'
+import { stringify } from '@/lib/utils/utils-shared'
 
+// For Overview of how timezones are handled, see docs/Implementation.md#timezones
+//
 // TODO: review what should be in this file vs date.ts
 
+const tzCache = new Map<string, boolean>()
+
 /**
- * Converts a UTC ISO 8601 wall time string to an ISO 8601 string
+ * Checks if a given timezone string is valid.
+ * isValidTimeZone('America/Los_Angeles') // true
+ * isValidTimeZone('UTC')                 // true
+ * isValidTimeZone('PST')                 // false
+ * isValidTimeZone('Mars/OlympusMons')    // false
+ * @param tz - Timezone string to validate (e.g., 'America/Los_Angeles')
+ * @returns True if the timezone is valid, false otherwise.
+ */
+export function isValidTimeZone(tz: string): boolean {
+    if (tzCache.has(tz)) return tzCache.get(tz)!
+
+    const valid = IANAZone.isValidZone(tz)
+    tzCache.set(tz, valid)
+    return valid
+}
+
+/**
+ * Reinterpret a UTC ISO 8601 wall time string to an ISO 8601 string
  * in a specified target timezone, preserving the wall time components.
+ * Or if input is a number (seconds since epoch), returns adjusted number (seconds since epoch)
  *
- * Example:
- *   '2024-06-10T17:00:00Z' + 'America/Los_Angeles'
- *   → '2024-06-10T17:00:00-07:00'
+ * Examples:
+ *   '2024-06-10T17:00:00Z','America/Los_Angeles' → '2024-06-10T17:00:00-07:00'
+ *   1718048400,'America/Los_Angeles' → 1718074800 ('2024-06-10T17:00:00-07:00')
  *
- * @param utcTime A date-time string in ISO 8601 format (e.g., 'YYYY-MM-DDTHH:mm:ssZ').
+ * @param utcTime Either number (seconds since epoch), or date-time string in ISO 8601 format (e.g., 'YYYY-MM-DDTHH:mm:ssZ').
  * The 'Z' indicates UTC, but its components (year, month, day, hour, minute, second)
  * are interpreted as the desired wall time to be preserved.
  * @param targetTimeZone The IANA timezone identifier (e.g., 'America/Los_Angeles', 'Europe/London').
  * @returns An ISO 8601 string in the format 'YYYY-MM-DDTHH:mm:ss+/-HH:mm'
  * with the original wall time components and the calculated offset for the target timezone.
  */
-export function convertUtcToTimeZone(utcTime: string, targetTimeZone: string): string {
-    return (
-        DateTime.fromISO(utcTime, { zone: 'utc' })
-            .setZone(targetTimeZone, { keepLocalTime: true })
-            .toISO({ suppressMilliseconds: true }) || ''
+export function reinterpretUtcTz(utcTime: string, targetTimeZone: string): string
+export function reinterpretUtcTz(utcTime: number, targetTimeZone: string): number
+export function reinterpretUtcTz(utcTime: string | number, targetTimeZone: string): string | number {
+    const fromUtc =
+        typeof utcTime === 'number'
+            ? DateTime.fromSeconds(utcTime, { zone: 'utc' })
+            : DateTime.fromISO(utcTime, { zone: 'utc' })
+
+    if (!fromUtc.isValid) {
+        throw new Error(`Invalid UTC time: ${utcTime}`)
+    }
+
+    // Reinterpret the wall time components as if they were in the target timezone
+    const reinterpreted = DateTime.fromObject(
+        {
+            year: fromUtc.year,
+            month: fromUtc.month,
+            day: fromUtc.day,
+            hour: fromUtc.hour,
+            minute: fromUtc.minute,
+            second: fromUtc.second,
+        },
+        { zone: targetTimeZone }
     )
+
+    if (reinterpreted && reinterpreted.isValid) {
+        return typeof utcTime === 'number'
+            ? Math.floor(reinterpreted.toSeconds())
+            : reinterpreted.toISO({ suppressMilliseconds: true })
+    } else {
+        throw new Error(`Invalid target timezone: ${targetTimeZone}`)
+    }
 }
 
 /**
- * Converts a UTC ISO 8601 string to an ISO 8601 string
- * in the timezone determined by latitude/longitude, preserving the wall time.
- *
- * Example:
- *   '2024-06-10T17:00:00Z' + (37.77, -122.42)  // San Francisco
- *   → '2024-06-10T17:00:00-07:00'
- * @param utcTime - The UTC ISO 8601 string format (e.g., 'YYYY-MM-DDTHH:mm:ssZ').
- * @param lat - Latitude
- * @param lon - Longitude
- * @returns An ISO 8601 string in the format 'YYYY-MM-DDTHH:mm:ss+/-HH:mm' or '' if invalid
+ * Convert a UTC ISO string (e.g. "2024-06-10T17:00:00Z")
+ * to epoch seconds (UTC).
  */
-export function convertUtcToTimeZoneAtCoords(utcTime: string, lat: number, lon: number): string {
-    const targetZone = tzlookup(lat, lon)
+export function convertUtcStringToSecs(utcIsoString: string): number {
+    const dt = DateTime.fromISO(utcIsoString, { zone: 'utc' })
+    if (!dt.isValid) {
+        throw new Error(`Invalid UTC ISO string: ${utcIsoString}`)
+    }
+    return Math.floor(dt.toSeconds())
+}
 
-    return (
-        DateTime.fromISO(utcTime, { zone: 'utc' })
-            .setZone(targetZone, { keepLocalTime: true })
-            .toISO({ suppressMilliseconds: true }) || ''
-    )
+/**
+ * Convert epoch seconds (UTC) to an ISO UTC string.
+ */
+export function convertUtcSecsToString(epochSeconds: number): string {
+    const dt = DateTime.fromSeconds(epochSeconds, { zone: 'utc' })
+    if (!dt.isValid) {
+        throw new Error(`Invalid epoch seconds: ${epochSeconds}`)
+    }
+    return dt.toISO({ suppressMilliseconds: true })
+}
+
+/**
+ * Update the time fields of events based on their location and timezone.
+ * This function is called by server /api/events, after fetching events and geocoding, before returning events
+ * @param events - Array of Event objects to update
+ * @returns Array of updated Event objects
+ */
+export function validateTzUpdateEventTimes(event: CmfEvent): CmfEvent {
+    if (event.tz === 'REINTERPRET_UTC_TO_LOCAL' && event.resolved_location) {
+        event.tz = event.resolved_location.location_tz
+        if (event.tz === 'UNKNOWN_TZ') {
+            logr.info('timezones', `Event with REINTERPRET_UTC_TO_LOCAL got UNKNOWN_TZ: ${stringify(event)}`)
+        } else if (event.tz && event.start && event.end) {
+            event.startSecs = reinterpretUtcTz(event.startSecs || convertUtcStringToSecs(event.start), event.tz)
+            event.endSecs = reinterpretUtcTz(event.endSecs || convertUtcStringToSecs(event.end), event.tz)
+            event.start = reinterpretUtcTz(event.start, event.tz)
+            event.end = reinterpretUtcTz(event.end, event.tz)
+            logr.info('timezones', `Event with REINTERPRET_UTC_TO_LOCAL set tz correctly: ${event.id}`)
+        } else {
+            throw new Error(`Event with REINTERPRET_UTC_TO_LOCAL is buggy: ${stringify(event)}`)
+        }
+    } else if (event.tz === 'REINTERPRET_UTC_TO_LOCAL' && !event.resolved_location) {
+        event.tz = 'UNKNOWN_TZ'
+        logr.info(
+            'timezones',
+            `Event with REINTERPRET_UTC_TO_LOCAL, no resolved_location: (bug) setting to UNKNOWN_TZ: ${stringify(event)}`
+        )
+    } else if (event.tz === 'TIME_IS_ACCURATE' && event.resolved_location) {
+        event.tz = event.resolved_location.location_tz
+        if (event.tz === 'UNKNOWN_TZ') {
+            event.tz === 'TIME_IS_ACCURATE' // keep it as is
+            logr.info('timezones', `Event with TIME_IS_ACCURATE, keeping,location_tz=UNKNOWN_TZ: ${stringify(event)}`)
+        } else {
+            // TODO: comment this out but leave for troubleshooting if need be
+            //logr.info('timezones', `Event with TIME_IS_ACCURATE, set tz correctly: ${event.id}`)
+        }
+    } else if (event.tz === 'TIME_IS_ACCURATE' && !event.resolved_location) {
+        event.tz = 'UNKNOWN_TZ'
+        logr.info(
+            'timezones',
+            `Event with TIME_IS_ACCURATE, no resolved_location: (bug) setting to UNKNOWN_TZ: ${stringify(event)}`
+        )
+    } else if (event.tz === 'UNKNOWN_TZ') {
+        throw new Error('Event with UNKNOWN_TZ before location lookup is a bug')
+    } else if (!event.tz) {
+        throw new Error('Event with no timezone is a bug - perhaps should have been REINTERPRET_UTC_TO_LOCAL')
+    } else if (isValidTimeZone(event.tz)) {
+        //logr.info('timezones', `GOOD event.tz='${event.tz}' for event: ${stringify(event)}`)
+    } else {
+        logr.info('timezones', `BAD event.tz='${event.tz}' for event: ${stringify(event)}`)
+    }
+    event.startSecs ||= convertUtcStringToSecs(event.start)
+    event.endSecs ||= convertUtcStringToSecs(event.end)
+    return event
 }
 
 /**
@@ -77,7 +184,7 @@ export function timezoneInfo(): { browserTz: string; tzOffset: string; tzAbbrev:
  * Gets the timezone for a given latitude and longitude
  * @param lat - Latitude
  * @param lng - Longitude
- * @returns Timezone string (e.g., 'America/Los_Angeles') or 'UNKNOWN_TZ' if not found
+ * @returns Timezone string (e.g., 'America/Los_Angeles') or 'UNKNOWN_TZ' if not found or error.
  */
 export const getTimezoneFromLatLng = (lat: number, lng: number): string => {
     try {